Minor cleanup

+++ src/menu.c
@@ -1141,7 +1141,7 @@ show_hidden_items(struct menu * menu, int force_clear)
                 STR_APPEND(hidden_msg, "%s", entry->name);
                 while (isspace(hidden_msg[strlen(hidden_msg)-1])) hidden_msg[strlen(hidden_msg)-1] = '\0';
                 while (ispunct(hidden_msg[strlen(hidden_msg)-1])) hidden_msg[strlen(hidden_msg)-1] = '\0';
-                hidden_msg[MIN(len+15, sizeof(hidden_msg))] = '\0';
+                hidden_msg[MIN(len+15, (int)sizeof(hidden_msg))] = '\0';
                 hidden_count++;
             }
             entry = entry->next;
